Generalize TANE and PFDTANE algorithms
Generalize TANE-based algorithms,
add additional pFD mining tests.
