+++ src/core/algorithms/fd/tane/model/lattice_vertex.h
