+++ src/core/config/error_measure/type.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "fd/pfdtane/enums.h"
+#include "fd/tane/enums.h"
 
 namespace config {
 using ErrorMeasureType = algos::ErrorMeasure;
